backend/requirements.txt
Flask==2.3.2
facenet-pytorch==2.6.0
opencv-python==4.10.0.84
numpy==1.25.0
pillow==10.2.0
flask-cors
pytz>=2020.1